+++ lib/src/command_runner.dart
@@ -1,5 +1,6 @@
 import 'package:args/command_runner.dart' as args;
 import 'package:dart_dependency_checker_cli/src/deps_unused/deps_unused_command.dart';
+import 'package:dart_dependency_checker_cli/src/transitive_use/transitive_use_command.dart';
 
 const versionFlag = 'version';
 
@@ -15,5 +16,6 @@ final class CommandRunner extends args.CommandRunner<int> {
       help: 'Print this package version.',
     );
     addCommand(DepsUnusedCommand());
+    addCommand(TransitiveUseCommand());
   }
 }